sl: update sorting library
git-svn-id: https://svn.version.fz-juelich.de/scafacos/trunk@2565 5161e1c8-67bf-11de-9fd5-51895aff932f

+++ lib/common/sl/extra/zmpi_atasp/zmpi_atasp.h
@@ -35,7 +35,7 @@
 #endif
 
 
-typedef int ZMPI_Count;
+typedef long ZMPI_Count;
 
 typedef struct _ZMPI_Tproc *ZMPI_Tproc;
